# contracts/openapi/aeon-echoes.v1.yaml
openapi: 3.1.0
info:
  title: Aeon Echoes HTTP API
  version: 1.0.0
  description: Versioned v1 API. All successful responses use the {data, meta} envelope; list responses also include page.
servers:
  - url: /api/v1
    description: v1 API root
paths:
...
  /projects/{projectID}/context-previews:
    parameters:
      - $ref: '#/components/parameters/ProjectID'
    post:
      operationId: previewContextSelection
      tags: [generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextPreviewRequest'
      responses:
        '200':
          description: Context preview.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/V1Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ContextPreviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{projectID}/chapters/{chapterID}/ideas:
    parameters:
      - $ref: '#/components/parameters/ProjectID'
      - $ref: '#/components/parameters/ChapterID'
    post:
      operationId: generateChapterIdea
      tags: [generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterIdeaRequest'
      responses:
        '201':
          description: Chapter idea generation result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/V1Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ChapterIdeaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{projectID}/chapters/{chapterID}/drafts:
    parameters:
      - $ref: '#/components/parameters/ProjectID'
      - $ref: '#/components/parameters/ChapterID'
    post:
      operationId: draftChapter
      tags: [generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftRequest'
      responses:
        '201':
          description: Draft result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/V1Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DraftResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{projectID}/character-profiles:
    parameters:
      - $ref: '#/components/parameters/ProjectID'
    post:
      operationId: generateCharacterProfiles
      tags: [generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterProfilesRequest'
      responses:
        '201':
          description: Character profile generation result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/V1Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CharacterProfilesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{projectID}/graph/expansions:
    parameters:
      - $ref: '#/components/parameters/ProjectID'
    post:
      operationId: expandGraph
      tags: [graph]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphExpansionRequest'
      responses:
        '200':
          description: Graph expansion.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/V1Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/GraphExpansion'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
...
    ProjectID:
      name: projectID
      in: path
      required: true
      schema:
        type: string
...
    ChapterID:
      name: chapterID
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    V1Envelope:
      type: object
      required: [data, meta]
      properties:
        data:
          description: Endpoint-specific payload.
        meta:
          $ref: '#/components/schemas/Meta'
...
    Meta:
      type: object
      required: [request_id]
      properties:
        request_id:
          type: string
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      required: [code, message, status, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        status:
          type: integer
        request_id:
          type: string
        details:
          description: Endpoint-specific error details.
...
    StoryBibleCharacter:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        desire:
          type: string
        wound:
          type: string
        secret:
          type: string
        summary:
          type: string
        entity_id:
          type: string
        sync_status:
          type: string
        synced_at:
          type: string
        metadata:
          $ref: '#/components/schemas/StringMap'
...
    CharacterProfileMapping:
      type: object
      properties:
        local_id:
          type: string
        name:
          type: string
        entity_id:
          type: string
        action:
          type: string
...
    ChapterVersion:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        chapter_id:
          type: string
        parent_version_id:
          type: string
        version:
          type: integer
        title:
          type: string
        content:
          type: string
        summary:
          type: string
        author_role:
          type: string
        source_workflow_id:
          type: string
        index_status:
          type: string
        metadata:
          $ref: '#/components/schemas/StringMap'
        created_at:
          type: string
          format: date-time
...
    ToolTrace:
      type: string
...
    IndexJob:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        chapter_id:
          type: string
        chapter_version_id:
          type: string
        kind:
          type: string
        status:
          type: string
        attempts:
          type: integer
        error:
          type: string
        payload:
          $ref: '#/components/schemas/StringMap'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        scheduled_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
    Workflow:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        kind:
          type: string
        role:
          type: string
        status:
          type: string
        model_id:
          type: string
        context_pack_id:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowStep'
        input:
          $ref: '#/components/schemas/StringMap'
        output:
          $ref: '#/components/schemas/StringMap'
        error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkflowStep:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        error:
          type: string
        metadata:
          $ref: '#/components/schemas/StringMap'
    Entity:
      type: object
      required: [id, project_id, name, type, summary, importance, status, created_at, updated_at]
      properties:
        id:
          type: string
        project_id:
          type: string
        worldline_id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [story_start, character, place, location, event, object, item, clue, chapter, concept, rule, time_node]
        aliases:
          type: array
          items:
            type: string
        summary:
          type: string
        traits:
          $ref: '#/components/schemas/StringMap'
        importance:
          type: integer
        status:
          type: string
          enum: [stable, draft, conflict, resolved, active, canonical, deprecated]
        metadata:
          $ref: '#/components/schemas/StringMap'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Fact:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        worldline_id:
          type: string
        entity_id:
          type: string
        chapter_id:
          type: string
        chapter_version_id:
          type: string
        claim:
          type: string
        source:
          type: string
        confidence:
          type: number
        status:
          type: string
        embedding_ref:
          type: string
        metadata:
          $ref: '#/components/schemas/StringMap'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GraphEdge:
      type: object
      required: [id, project_id, source_entity_id, target_entity_id, type, label, weight, created_at, updated_at]
      properties:
        id:
          type: string
        project_id:
          type: string
        worldline_id:
          type: string
        source_entity_id:
          type: string
          minLength: 1
        target_entity_id:
          type: string
          minLength: 1
        type:
          type: string
        label:
          type: string
        weight:
          type: number
        evidence_fact_ids:
          type: array
          items:
            type: string
        metadata:
          $ref: '#/components/schemas/StringMap'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PlotThread:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        worldline_id:
          type: string
        title:
          type: string
        summary:
          type: string
        status:
          type: string
        priority:
          type: integer
        related_entity_ids:
          type: array
          items:
            type: string
        opened_chapter_id:
          type: string
        closed_chapter_id:
          type: string
        metadata:
          $ref: '#/components/schemas/StringMap'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GraphExpansion:
      type: object
      required: [project_id, depth, entities, edges, facts, generated_at]
      properties:
        project_id:
          type: string
        depth:
          type: integer
          minimum: 1
          maximum: 4
        entities:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'
        facts:
          type: array
          items:
            $ref: '#/components/schemas/Fact'
        generated_at:
          type: string
          format: date-time
...
    ModelResolution:
      type: object
      properties:
        route_key:
          type: string
        resolution_source:
          type: string
        provider_id:
          type: string
        provider_name:
          type: string
        provider_type:
          type: string
        model_id:
          type: string
        model_name:
          type: string
        model_kind:
          type: string
    IndexFreshness:
      type: object
      properties:
        project_id:
          type: string
        chapter_id:
          type: string
        status:
          type: string
        latest_chapter_version_id:
          type: string
        latest_chapter_version_created_at:
          type: string
          format: date-time
        latest_indexed_chapter_version_id:
          type: string
        latest_indexed_at:
          type: string
          format: date-time
        pending_job_count:
          type: integer
    ContinuityEvidenceRef:
      type: object
      properties:
        source_type:
          type: string
        source_id:
          type: string
        label:
          type: string
        excerpt:
          type: string
    ContinuityIssue:
      type: object
      properties:
        type:
          type: string
        severity:
          type: string
        message:
          type: string
        draft_excerpt:
          type: string
        suggestion:
          type: string
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/ContinuityEvidenceRef'
    ContinuityAudit:
      type: object
      properties:
        status:
          type: string
        issues:
          type: array
          items:
            $ref: '#/components/schemas/ContinuityIssue'
    ChapterSummary:
      type: object
      properties:
        chapter_id:
          type: string
        chapter_version_id:
          type: string
        title:
          type: string
        summary:
          type: string
    ContextPack:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        chapter_id:
          type: string
        role:
          type: string
        token_budget:
          type: integer
        query:
          type: string
        story_bible_id:
          type: string
        world_rules:
          $ref: '#/components/schemas/StringMap'
        facts:
          type: array
          items:
            $ref: '#/components/schemas/Fact'
        entities:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'
        plot_threads:
          type: array
          items:
            $ref: '#/components/schemas/PlotThread'
        chapter_summaries:
          type: array
          items:
            $ref: '#/components/schemas/ChapterSummary'
        tool_trace:
          type: array
          items:
            $ref: '#/components/schemas/ToolTrace'
        metadata:
          $ref: '#/components/schemas/StringMap'
        created_at:
          type: string
          format: date-time
    ContextPreviewResponse:
      type: object
      properties:
        context_pack:
          $ref: '#/components/schemas/ContextPack'
        summary:
          type: string
        estimated_tokens:
          type: integer
        index_freshness:
          $ref: '#/components/schemas/IndexFreshness'
        model_resolution:
          $ref: '#/components/schemas/ModelResolution'
    ChapterIdeaResponse:
      type: object
      properties:
        workflow:
          $ref: '#/components/schemas/Workflow'
        context_pack:
          $ref: '#/components/schemas/ContextPack'
        chapter_idea:
          type: string
        model_resolution:
          $ref: '#/components/schemas/ModelResolution'
        tool_trace:
          type: array
          items:
            type: string
    CharacterProfilesResponse:
      type: object
      properties:
        workflow:
          $ref: '#/components/schemas/Workflow'
        context_pack:
          $ref: '#/components/schemas/ContextPack'
        characters:
          type: array
          items:
            $ref: '#/components/schemas/StoryBibleCharacter'
        entities:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        mappings:
          type: array
          items:
            $ref: '#/components/schemas/CharacterProfileMapping'
        model_resolution:
          $ref: '#/components/schemas/ModelResolution'
        tool_trace:
          type: array
          items:
            type: string
    DraftResponse:
      type: object
      properties:
        workflow:
          $ref: '#/components/schemas/Workflow'
        context_pack:
          $ref: '#/components/schemas/ContextPack'
        chapter_version:
          $ref: '#/components/schemas/ChapterVersion'
        index_job:
          $ref: '#/components/schemas/IndexJob'
        index_freshness:
          $ref: '#/components/schemas/IndexFreshness'
        model_resolution:
          $ref: '#/components/schemas/ModelResolution'
        continuity_audit:
          $ref: '#/components/schemas/ContinuityAudit'
        tool_trace:
          type: array
          items:
            type: string
    ContextSelection:
      type: object
      additionalProperties: false
      properties:
        chapter_ids:
          type: array
          items:
            type: string
        character_ids:
          type: array
          items:
            type: string
        character_names:
          type: array
          items:
            type: string
        include_world_rules:
          type: boolean
    ContextPreviewRequest:
      type: object
      additionalProperties: false
      properties:
        chapter_id:
          type: string
        title:
          type: string
        brief:
          type: string
        prompt:
          type: string
        context_node_ids:
          type: array
          items:
            type: string
        context_selection:
          $ref: '#/components/schemas/ContextSelection'
        reference_selection:
          $ref: '#/components/schemas/ContextSelection'
        style_constraints:
          type: array
          items:
            type: string
        role:
          type: string
        token_budget:
          type: integer
    ChapterIdeaRequest:
      allOf:
        - $ref: '#/components/schemas/ContextPreviewRequest'
        - type: object
          properties:
            max_output_tokens:
              type: integer
    DraftRequest:
      allOf:
        - $ref: '#/components/schemas/ChapterIdeaRequest'
        - type: object
          properties:
            chapter_idea:
              type: string
            chapter_idea_workflow_id:
              type: string
    CharacterProfilesRequest:
      type: object
      additionalProperties: false
      properties:
        focus:
          type: string
        count:
          type: integer
        brief:
          type: string
        chapter_id:
          type: string
        context_node_ids:
          type: array
          items:
            type: string
        context_selection:
          $ref: '#/components/schemas/ContextSelection'
        max_output_tokens:
          type: integer
    GraphExpansionRequest:
      type: object
      additionalProperties: false
      required: [depth]
      properties:
        entity_ids:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
        depth:
          type: integer
          minimum: 1
          maximum: 4
...
    StringMap:
      type: object
      additionalProperties:
        type: string
...